fix(server): bound the shutdown announcement by the drain budget
The drain announced to every session and awaited it with no timeout, fanning out unbounded. A client that has stopped reading TCP therefore held the announcement for the whole per-send SendTimeout — 30s by default, against a 5s drain budget and a 15s host budget — which is long enough for rask deploy's SIGKILL to land in the middle of a SQLite checkpoint.

The announcement now goes through LiveSessionStore.BroadcastAsync (bounded concurrency, faults skipped) and is awaited under the budget token like every other drain step.

Only visible once the drain and the bounded outbound send were on the same base: each was written against a tree without the other, and the drain's own comment still claimed BroadcastAsync was sequential, which stopped being true when the send bound landed.

Measured by the new test: 30s before, 200ms after.
